Renamed package from guide to guides
