+++ src/main/java/co/cask/cdap/guides/PageRankApplication.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package co.cask.cdap.guide.pagerank;
+package co.cask.cdap.guides;
 
 import co.cask.cdap.api.app.AbstractApplication;
 import co.cask.cdap.api.dataset.lib.ObjectStores;
